apps/api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
psycopg2-binary>=2.9
pydantic>=2.0
pydantic-settings>=2.0
alembic>=1.13
python-multipart>=0.0.9
numpy
opencv-python-headless
Pillow
PyMuPDF>=1.24
pypdf
